+++ src/GridCal/Engine/Simulations/Stochastic/lhs_driver.py
@@ -145,7 +145,6 @@ class LatinHypercubeSampling(QThread):
                     # store circuit results at the time index 't'
                     lhs_results.S_points[t, numerical_island.original_bus_idx] = res.Sbus
                     lhs_results.V_points[t, numerical_island.original_bus_idx] = res.voltage
-                    lhs_results.I_points[t, numerical_island.original_branch_idx] = res.Ibranch
                     lhs_results.Sbr_points[t, numerical_island.original_branch_idx] = res.Sbranch
                     lhs_results.loading_points[t, numerical_island.original_branch_idx] = res.loading
                     lhs_results.losses_points[t, numerical_island.original_branch_idx] = res.losses
@@ -233,7 +232,6 @@ class LatinHypercubeSampling(QThread):
                     # Gather the results
                     lhs_results.S_points[t, numerical_island.original_bus_idx] = S
                     lhs_results.V_points[t, numerical_island.original_bus_idx] = res.voltage
-                    lhs_results.I_points[t, numerical_island.original_branch_idx] = res.Ibranch
                     lhs_results.Sbr_points[t, numerical_island.original_branch_idx] = res.Sbranch
                     lhs_results.loading_points[t, numerical_island.original_branch_idx] = res.loading
                     lhs_results.losses_points[t, numerical_island.original_branch_idx] = res.losses