+++ src/GridCal/Engine/Simulations/Stochastic/monte_carlo_driver.py
@@ -186,7 +186,7 @@ class MonteCarlo(QThread):
                         # store circuit results at the time index 't'
                         mc_results.S_points[t, numerical_island.original_bus_idx] = res.Sbus
                         mc_results.V_points[t, numerical_island.original_bus_idx] = res.voltage
-                        mc_results.I_points[t, numerical_island.original_branch_idx] = res.Ibranch
+                        mc_results.Sbr_points[t, numerical_island.original_branch_idx] = res.Sbranch
                         mc_results.loading_points[t, numerical_island.original_branch_idx] = res.loading
                         mc_results.losses_points[t, numerical_island.original_branch_idx] = res.losses
 
@@ -301,7 +301,7 @@ class MonteCarlo(QThread):
 
                         mc_results.S_points[t, numerical_island.original_bus_idx] = res.Sbus
                         mc_results.V_points[t, numerical_island.original_bus_idx] = res.voltage
-                        mc_results.I_points[t, numerical_island.original_branch_idx] = res.Ibranch
+                        mc_results.Sbr_points[t, numerical_island.original_branch_idx] = res.Sbranch
                         mc_results.loading_points[t, numerical_island.original_branch_idx] = res.loading
                         mc_results.losses_points[t, numerical_island.original_branch_idx] = res.losses
 
@@ -338,12 +338,8 @@ class MonteCarlo(QThread):
                 std_dev_progress = 100
             self.progress_signal.emit(max((std_dev_progress, it / self.max_mc_iter * 100)))
 
-            # print(iter, '/', max_mc_iter)
-            # print('Vmc:', Vavg)
-            # print('Vstd:', Vvariance, ' -> ', std_dev_progress, ' %')
-
         # compile results
-        mc_results.sbranch = avg_res.Sbranch
+        # mc_results.sbranch = avg_res.Sbranch
         # mc_results.losses = avg_res.losses
         mc_results.bus_types = numerical_circuit.bus_types